arm64: dts: ti: k3-am64-main: Add hwspinlock node
The AM64x SoCs contains a HwSpinlock IP instance that is based on the
IP on AM65x SoCs. Add the DT node for this on AM64x SoCs. The node is
present within the the Main domain, and is added as a child node under
the cbass_main node.

Signed-off-by: Suman Anna <s-anna@ti.com>