TEST: hwspinlock/omap: Add a stand-alone unit-test module
Add an unit-test module to test all the possible hwspinlocks supported
on a bank using the regular hwspinlock request API. The module also
tests DT client usage optionally (based on presence of a test-node)
using the appropriate OF request API of_hwspin_get_lock_id() request
API.

The test module is designed as a regular module instead of a platform
driver so that it can be run even without the presence of any test nodes.
The module uses some SoC data for figuring out the maximum
number of locks on a SoC.

The code is based on the stand-alone kernel unit-test repository staged
at https://github.com/sumananna/omap-hwspinlock-test, and updated for
the AM64 SoC detection.

Signed-off-by: Suman Anna <s-anna@ti.com>