TST: arm64: dts: ti: k3-am64-zebu: Add mailbox loopback devices for testing
Add some loopback sub-mailbox devices within each of the Mailbox Clusters
2 through 7 on the AM64 SoCs for testing on Zebu. The loopback device uses
the same mailbox fifo for sending and receiving messages on the MPU itself,
facilitating a unit-test.

Two sub-mailbox devices were added on clusters having 2 interrupt lines
connected to the A53 GIC to exercise each of the interrupt line. Different
FIFOs were also chosen with the second sub-mailbox device in the applicable
clusters, and not to conflict with the actual FIFOs at the beginning that
will be used to communicate with the remote processors.

Signed-off-by: Suman Anna <s-anna@ti.com>