dmaengine: ti: k3-udma: Initial support for KSLC PKTDMA
One of the DMAs in KSLC architecture is the Packet DMA (PKTDMA).
It offers similar features as K3 UDMAP channels in packet mode.

PKTDMA can be used to service PSI-L peripherals, similarly to
K3 UDMA channels.

Most of the driver code can be reused for PKTDMA tchan/rchan support but
new setup and allocation functions are needed to handle the differences
between the DMAs.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>