dmaengine: ti: k3-udma-glue: Use k3_ringacc_request_rings_pair() API
On K3 architectures, DMA descriptor submission and completion rings are
separate whereas they are one and the same on KSLC architecture.
Therefore convert glue layer to use the new k3_ringacc_request_rings_pair()
API to request rings, as this API takes care of abstracting both the types.

This is in preparation to add support for KSLC DMA

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>