HACK: VLAB: net: ethernet: ti: am65-cpsw: reduce num descs
AM64x VLAB has an issue which does not allow to push >127 descriptors in to
the ring as OCC is updated incorrectly, like:

OCC = OCC % 128.

It was also observed that Port_id=0 returned in RX descriptor sometimes
under have traffic, so protect our-self from crash. This may be related to
above ring issue.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>