dmaengine: ti: k3-udma: Support for per channel coherency for am64
KSLC DMA architecture allows selection of coherency by channels. This means
that channels individually can be configured to be considered coherent
or non coherent.

The coherency is selected via ASEL (bits 48-51) part of the 64 bits address
region.

If a channel's ASEL is 14 or 15, then the DMA access are routed via ACP
port to achieve coherency.

Update the channel's device if it is configured to be coherent and let
users to get the channel's device for DMA API use via the
dmaengine_get_device_for_dma_api() call.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>