arm64: dts: ti: k3-am64-vlab: Rename secure_ddr reserved-memory node
The AM64 VLAB reserves a memory region for OPTEE using the name
secure_ddr. Underscores should not be used in node names. Fix this
by renaming the "secure_ddr" node name with "optee" as used on J721E
and J7200 board dts files.

The license headers have also been updated to include 2020 in the
copyright year and to use the https reference links.

Fixes: 1ac93d54c0fa ("arm64: boot: dts: ti: vlab: Add AM64 support")
Signed-off-by: Suman Anna <s-anna@ti.com>