arm64: dts: ti: k3-am64-vlab: Move R5F Core1 sub-mailbox nodes
Commit d7e1a916505b ("TEMP: arm64: dts: ti: k3-am64-vlab: Add IPC
sub-mailbox nodes") has added the sub-mailboxes for each of the R5F
cores in a unique cluster. Move the sub-mailboxes associated with
each of the R5F Core1 to the same mailbox cluster as the corresponding
R5F Core0 to match the final usage and aligned with the latest
firmwares in TI RTOS PDK project.

This change will essentially result in using Mailbox Cluster 2 for
Main R5FSS0, Cluster 4 for Main R5FSS1 and Cluster 6 for M4F.

Signed-off-by: Suman Anna <s-anna@ti.com>