arm64: dts: ti: k3-am64-zebu: Reserve memory for IPC between RTOS cores
Add a reserved memory node to reserve a portion of the DDR memory to be
used for performing inter-processor communication between all the remote
processors running RTOS on the TI K3 AM64 Zebu platform. 8 MB of memory is
reserved for this purpose, and this accounts for all the vrings and vring
buffers between all the possible pairs of remote processors.

NOTE:
The AM64x SoCs do not have any DSPs and one less R5F cluster compared
to J721E SoCs. So, while the carveout memories reserved for the R5F
clusters present on the SoC match to those on J721E, the overall memory
map reserved for firmwares is quite different. The number of R5F
clusters on AM64x SoCs are same as on J7200 SoCs, but the AM64x SoCs
also have an additional M4F core, so the RTOS IPC memory region is
1 MB higher than on J7200 SoCs.

Signed-off-by: Suman Anna <s-anna@ti.com>