HACK: remoteproc: k3-r5: Hardcode Split-mode for AM64x SoCs
The K3 AM64x SoC family has a revised R5F sub-system that supports
two modes - a new "Single-CPU" mode and the conventional Split mode.
Some devices in the SoC family support only a "Single-CPU" mode, but
the default mode on most applicable devices is Split-mode.

The Single-CPU mode is not currently supported in the driver, the
power-up and power-down logic for this mode needs to be vetted. So,
hard-code the mode to the default Split-mode in case the Single-CPU
mode is selected in the dts nodes or board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>