dmaengine: ti: k3-udma: Fix rebase issues
Seems that some rebase error has caused a second of_match_node to be
introduced in the probe function that causes probe to fail, so this must
be dropped.

Also, the AM64 SoC ID used in matching in AM64, so correct the AM64X
that was introduced somewhere for this driver.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>