arm64: dts: ti: k3-am64-base-board: Add mailboxes to R5Fs
Add the required 'mboxes' property to all the R5F processors on the
TI AM64x EVM board. The mailboxes and some shared memory are required
for running the Remote Processor Messaging (RPMsg) stack between the host
processor and each of the R5Fs. The chosen sub-mailboxes match the values
used in the current firmware images. This can be changed, if needed, as
per the system integration needs after making appropriate changes on the
firmware side as well.

Note that any R5F Core1 resources are needed and used only when that
R5F cluster is configured for Split-mode.

Signed-off-by: Suman Anna <s-anna@ti.com>