remoteproc: k3-r5: Fix support for Single-CPU mode on AM64x SoCs
The K3 AM64x SoC family has a revised R5F sub-system that supports
two modes - a new "Single-CPU" mode and the conventional Split mode.
The Single-CPU mode is somewhat similar to the LockStep-mode on J7200
SoCs.

The mode does not provide fault-tolerance like LockStep-mode, executes
code only on a single CPU (on Core0), but uses the TCMs from both the
R5F cores. As such, this requires the resets to be managed for both
the cores, but only the halt signal needs to be programmed for Core0.
The reset

Support for this Single-CPU mode was not enabled previously. Fix the
driver to handle this mode properly. The reset release sequence is
agnostic of the order, so the same LockStep reset and release sequences
are re-used for minimal code.

Signed-off-by: Suman Anna <s-anna@ti.com>