Revert "HACK: remoteproc: k3-r5: Hardcode Split-mode for AM64x SoCs"
This reverts commit 84368acbfbe32e4aead2488e77c011c73771abff.

The Single-CPU mode is not supported fully in the driver previously,
and the mode was hard-coded by default to Split-mode in the probe
function in case the Single-CPU mode was selected in the dts nodes
or board dts files.

This has since been fixed in commit ea3cbbc4904a ("remoteproc: k3-r5:
Fix support for Single-CPU mode on AM64x SoCs"), so revert this
hard-coded logic.

Signed-off-by: Suman Anna <s-anna@ti.com>