arm64: dts: ti: k3-j721s2-som-p0: Add DDR carveout memory nodes and assign mailboxes for R5Fs and C71xs
Two carveout reserved memory nodes each have been added for each of the
R5F and C71x remote processor devices within both the MCU and MAIN domains
on the TI J721S2 SoCs. These nodes are assigned to the respective
rproc device nodes as well. The first region will be used as the DMA pool
for the rproc device, and the second region will furnish the static
carveout regions for the firmware memory.

An additional reserved memory node is also added to reserve a portion of
the DDR memory to be used for performing inter-processor communication
between all the remote processors running RTOS. 28 MB of memory is
reserved for this purpose, and this accounts for all the vrings and vring
buffers between all the possible pairs of remote processors.

The current carveout addresses and sizes are defined statically for each
device. The R5F processors do not have an MMU and C71x does support a MMU
called CMMU, but is not currently supported. Therefore, they require
the exact memory used by the firmwares to be set-aside. The firmware images
do not require any RSC_CARVEOUT entries in their resource tables either
to allocate the memory for firmware memory segments.

Therefore, add the DDR carverout memories and assign mailboxes to the
remoteprocs in their respective device tree nodes.

Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>