arm64: dts: k3-j721s2: Hog gpio expander pins for selecting CAN signal in muxes 1 and 2
GPIO expander pins 1,2 and 3 are used to select the output lines in
SoM to CPB muxes 1 and 2. These muxes control connections of the Rx/Tx
lines of main_mcan3 and main_mcan5 to CPB's MCAN Tx/Rx lines of the phy
transceiver. Therefore, hog the output to high for these gpios pins.

Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>