arm64: dts: ti: k3-am625-sk: Add nodes to enable display
Add interrupt and pin mux details for the IO expander chip and the pin
mux for the mcu_gpio pin that controls the interrupt. The IO expander
acts as an interrupt controller for the sil9022 hdmi transmitter.

Add pin mux details in the dss node of the AM625-SK.

Add the hdmi transmitter node details on the AM625-SK board, and its
master clock. The hdmi transmitter used on AM625-SK is sii9022 and is
connected on the i2c1 bus.

Add hdmi-connector node, its input port, input and output ports of the
sii9022 hdmi transmitter, and the DPI1 output port of dss on the
AM625-SK. The DPI1 output of DSS connects to the input of sii9022, and
the output of sii9022 connects to the input of the hdmi-connector.

Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>