arm64: dts: ti: k3-j784s4-evm: Speed up sd/mmc
Lets go from 1 bit mode to 4 bit mode, but without 1v8, we will not have
HS200 mode working. lets keep that disabled for now.

ALMOST ALWAYS, HS200 tends to fail on the wide majority of samples
until complete characterization is complete, so it might be good idea to
just disable it at the root node itself.

Signed-off-by: Nishanth Menon <nm@ti.com>