firmware: ti_sci: Introduce prepare system suspend call
Upstream:
https://lore.kernel.org/all/20221128140522.49474-7-g-vlaev@ti.com/

Introduce a ti_sci_prepare_system_suspend call to be used in the driver
suspend handler to allow the system to identify the low power mode being
entered and if necessary, send TISCI_MSG_PREPARE_SLEEP with information
about the mode is being entered and the address for allocated memory for
storing the context during Deep Sleep.

We're using "pm_suspend_target_state" to map the kernel's target suspend
state to SysFW low power mode. Make sure this is available only when
CONFIG_SUSPEND is enabled.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Georgi Vlaev <g-vlaev@ti.com>
[backport: Selected relevant hunks for ti-linux-5.10.y]
Signed-off-by: Vibhore Vardhan <vibhore@ti.com>