HACK: drm/panel-edp: Fix atomic check errors for panel
When we do the tidss_encoder_atomic_check, if the connector does
not have the bus_format, the check fails and we get EINVAL.
This bus_format object was removed from panel-edp driver when
it was split from panle-simple driver. So add the bus_format
back to the panel_desc structure and set it in connector
display_info.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>