[HACK] dmaengine: ti: k3-udma: Force terminate peer for McASP RX
On AM62 and AM62a, when McASP PDMA is the peer, doing a normal teardown
on the RX channel can often lead to channel corruption. In such cases it
is impossible to recover the channel without resetting the SoC, and the
corrupt channel receives no data and can even cause CPU lockups.

Instead if we do a forced disable on the peer endpoint (which requires a
channel reset and PSI-L re-pairing) we can avoid the channel corruption,
at the cost of doing a hard reset after every recording.

Signed-off-by: Jai Luthra <j-luthra@ti.com>