media: ti: j721e-csi2rx: Fix buffer cleanup sequence
Ensure that DMA is terminated (and any active buffers are marked as done
in the dma channel callback) before marking the remaining buffers in the
queue as done.

Without this fix there were rare errors reported inside dma callback if
the active buffer was already returned to VB2 and thus got freed twice.

Fixes: 63bc1fbc5c85 ("media: ti: j721e-csi2rx: Fix stream stop sequence")
Signed-off-by: Jai Luthra <j-luthra@ti.com>