media: ti: j721e-csi2rx: Fix broken UYVY format ordering
There was an incorrect assumption in the driver that we need to set a
different YUV422_MODE in SHIM hardware to support different incoming
UYVY/YUYV/.. variants.

The hardware in-fact assumes incoming data is UYVY following MIPI CSI
spec, and the YUV422_MODE is available in case we want to swap the bytes
around before storing in memory.

Given some sensors (like OV5640) can send both UYVY and YUYV data, and
we don't have a requirement of swapping them to something else, we
should simply "passthrough" the incoming packed format by using MODE_11.

Reported-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Fixes: 70f7e7ba6c66 ("media: ti: Add CSI2RX support for J721E")
Signed-off-by: Jai Luthra <j-luthra@ti.com>