[uL][RFC] CSI: Link Cadence & TI CSI nodes using graph endpoints
This is an internal RFC for fixing the missing link between Cadence
CSI2RX and TI Shim IPs.

IMHO, as of today the Cadence CSI2RX node is modelled wrongly in the
device tree. The Port 1 (corresponding to Cadence IP's stream0) is
disabled, when in the hardware it is connected to the Shim block.

This issue was found during an upstream discussion in the v2 of multistream
series:
https://lore.kernel.org/all/x3hku4b4ybomt3jttwckggvagzasfnj3bdztsiei5uedj5xu4n@go5lpd6kugdf/

The fixes to the DTS files need to be present before the driver change
introducing the .get_fwnode_pad callback.

Nishanth, Vignesh,
will it possible to merge dt-bindings (through media tree) and DT node
changes (through ti tree) in the next cycle, and the driver
change (through media tree) in the next-to-next cycle?

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

Signed-off-by: Jai Luthra <j-luthra@ti.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240911-csi_endpoint_fix-21a49da39911",
    "prefixes": []
  }
}