media: dt-bindings: ti,ds90ub960: Add "i2c-addr" link property
The serializer's I2C address on the FPD-Link bus is usually communicated
to the deserializer once the forward-channel is established. But in some
cases it might be necessary to program the serializer (over the
back-channel) before the forward-channel is established.

Add a new optional property to specify the I2C address of the
serializer.

Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>