media: i2c: imx219: Increase minimum LLP for binned modes
The sensor's internal ADC supports a minimum line length of 3448 pixels,
which may be too small to use with analog binning, where ADC operates on
two lines together. Switch to a higher minimum line length of 3560
pixels to fix the blocky artefacts seen with analog binning.

To keep the same default framerate as before, lower the default fll value
for the binned modes to compensate for the increase in llp.

Link: https://github.com/raspberrypi/rpicam-apps/issues/281#issuecomment-1082894118
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>