media: i2c: imx219: Only use higher LLP_MIN for binned resolutions
The LLP_MIN of 3560 is only needed to fix artefacts seen with binned
resolutions. As increasing the LLP reduces the highest possible
framerate by ~3%, use the old default minimum of 3448 if we are not
doing 2x2 analog binning.

Also restore the fll_def value for non-binned modes in the modes
definition to restore the default mode framerate to 30fps.

Suggested-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Link: https://lore.kernel.org/linux-media/CAPY8ntC1-S6zKtDvmc6EgyxP+j6rTShuG8Dr8PKb9XQr2PeS_w@mail.gmail.com/
Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>
---
Changes in v2:
    - Fix the binning check, and only update the LLP minimum if we are
      doing analog binning
    - Add a note in the commit message for the fll_def value reverts