media: ti: j721e-csi2rx: Fix source subdev link creation
We don't currently use OF ports and remote-endpoints to connect the
CSI2RX bridge and this device in the device tree, thus it is wrong to
use v4l2_create_fwnode_links_to_pad() to create the media graph link
between the two.

It works out on accident, because neither the source nor the sink
implement the .get_fwnode_pad() callback so the framework helper falls
back on using the first source and sink pads between them.

Instead, manually create the media link from the first source pad of the
bridge, as this connection is assumed by the device tree.

Fixes: b4a3d877dc92 ("media: ti: Add CSI2RX support for J721E")
Cc: stable@vger.kernel.org
Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>