TO-BE-SPLIT: Connect cdns and j721e csi2rx using OF graph
This adds support for using OF graph endpoint to represent the
connection between port@1 of cdns-csi2rx and j721e-csi2rx. Earlier this
was not modelled in the device tree, but hardcoded in the j721e-csi2rx
driver.

On TI SoCs that have an ISP, the j721e-csi2rx IP supports ingesting an
additional pixel stream from port@3 of cdns-csi2rx that it forwards to the
ISP directly instead of going through DDR. It is cleaner to be explicit
about such connections in the device tree rather than hardcoding them in
the drivers.

But to maintain backwards compatibility with existing DT firmware, the
ports definition is kept optional in the bindings, and without it the
driver would fallback to using the hardcoded values.