media: v4l2-ioctl: Pass device state for G/S/TRY_FMT ioctls
Now that video device state is accepted as an argument in ioctl
implementations across all video device drivers, populate that argument
with the actual state data.

Pass the active state for G_FMT and S_FMT ioctls, and the try state for
TRY_FMT ioctl to drivers that have opted to use device state for format
storage.

Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>
--
Cc: Mauro Carvalho Chehab <mchehab@kernel.org>
Cc: Hans Verkuil <hverkuil@kernel.org>
Cc: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Cc: Ricardo Ribalda <ribalda@chromium.org>
Cc: Hans de Goede <hansg@kernel.org>
Cc: Jai Luthra <jai.luthra@ideasonboard.com>
Cc: linux-media@vger.kernel.org
Cc: linux-kernel@vger.kernel.org