media: rkisp1: Calculate format information on demand
Instead of storing format configuration and information as explicit
state in the driver structure, calculate it on demand when needed using
the format stored in the active video device state.

This change removes the pix member from rkisp1_capture structure and
refactors the code to look up format configuration and v4l2_format_info
at the point of use.

Additionally, grey format handling is moved from the next buffer setup
path (which may be called in interrupt context) to the buffer
initialization phase for simplicity.

Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>
---
Cc: Dafna Hirschfeld <dafna@fastmail.com>
Cc: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Cc: Mauro Carvalho Chehab <mchehab@kernel.org>
Cc: Heiko Stuebner <heiko@sntech.de>
Cc: linux-media@vger.kernel.org
Cc: linux-rockchip@lists.infradead.org
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-kernel@vger.kernel.org