media: ipu6: Bridge the gap between streams in V4L2 and IPU6 firmware
The stream concept of the IPU6 firmware is aligned with that of the CSI-2,
effectively meaning that a stream equals to a CSI-2 virtual channel. Only
start streaming once all streams on a given virtual channel have been
started, and stop streaming once first one of them has been stopped.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>